Time: 373 ms (41.67%), Space: 107.9 MB (93.92%) - LeetHub

+++ 1524-number-of-sub-arrays-with-odd-sum/1524-number-of-sub-arrays-with-odd-sum.cpp
@@ -5,8 +5,8 @@ public:
         int ans = 0;
         int val[2] = {0};
         val[0] = 1;
-        int currSum = 0;
-        for(int num: arr) {
+        short currSum = 0;
+        for(short num: arr) {
             currSum = (currSum + num) % 2;
             ans = (ans + val[1 - currSum]) % MOD;
             val[currSum]++;
